Move objdet functions to a separated C++ file
